+++ python/create-onchain-agent/CHANGELOG.md
@@ -1,3 +1,14 @@
+# Coinbase Create Onchain Agent Changelog
+
+<!-- towncrier release notes start -->
+
+## [0.4.0] - 2025-03-28
+
+### Added
+
+- Updated default wallet provider to SmartWalletProvider ([#622](https://github.com/coinbase/agentkit/pull/622))
+
+
 ## [0.3.0] - 2025-03-14
 
 ### Added
@@ -17,8 +28,6 @@
 - Added revised network selection ([#498](https://github.com/coinbase/agentkit/pull/498))
 - Added smart wallet support ([#503](https://github.com/coinbase/agentkit/pull/503))
 
-
-# Coinbase Create Onchain Agent Changelog
 
 ## [0.1.4] - 2025-02-28