+++ python/framework-extensions/openai-agents-sdk/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Coinbase AgentKit Framework Extension - OpenAI Agents SDK Changelog
 
+<!-- towncrier release notes start -->
+
+## [0.2.0] - 2025-03-28
+
+### Added
+
+- Bump coinbase-agentkit to 0.2.0 ([#608](https://github.com/coinbase/agentkit/pull/608))
+- Bump openai-agents to v0.0.6 ([#609](https://github.com/coinbase/agentkit/pull/609))
+
+
 ## [0.1.0] - 2025-03-11
 
 ### Added